di5cheng-UT-Fleet-P/requirements.txt
Appium-Python-Client==0.49
selenium==3.141.0
pytest==5.3.2
PyYAML==5.2
allure-pytest==2.8.6
python-docx==0.8.10
openpyxl==3.0.2
xlrd==1.2.0
xlwt==1.3.0
